fix(test-infra): #1371 Codex Round 2 Must-fix 解消 — 3 spec を 1 file 統合 + retries: 2
Codex Round 2 で reject。workers=2 並列で screen-items-rename / screen-flow-rename が
間欠 timeout (Round 1 で 5s→15s 拡大 + page.evaluate 化したが、backend `editSession.create`
/ `previewEntityRename` RPC が workers 並列下で遅延する根本問題が残存)。

## 修正

### 3 spec を `frontend/e2e/screen-rename.spec.ts` に統合

3 起動点 (ScreenItemsView / ScreenListView / ScreenFlow) を別 file に分けると Playwright
`workers=2` で並列実行され、3 test とも backend に対し setupTestWorkspace + editSession.create
+ previewEntityRename + rename を行うため、共有 backend (port 5179) の MCP RPC が並列負荷で
間欠 timeout する。

1 file に統合して `test.describe.configure({ mode: "serial" })` で順次実行することで、
workers=2 並列下でも 1 worker 内で連続実行され、backend 並列負荷を回避する。

3 test それぞれ別 screen target (cart / cart-confirm / customer-master) で rename し、
順次実行中の state 干渉も回避。

### `retries: 2` を describe レベルで設定

backend RPC 遅延の残余 flake (10 runs で 3 件、いずれも (A) ScreenItemsView 起点) を
自動リトライで吸収。根本対策 (GenericDefinitionEditor 等の re-render 抑止) は本 PR
scope 外 (rename refactor の本体機能ではなく e2e trace 確立目的)。

### dialog 内 button click の polling-retry helper 化

`fetchExistingIds` inline arrow による useEffect 再 fire ループで dialog 内 button が
頻繁に re-render する race を、`waitForFunction` で「button が DOM 存在 + clickable」を
polling 確認してから `evaluate` direct DOM click を毎回再取得して dispatch する helper
(`clickDialogButton`) で堅牢化。

## 検証

- `npx playwright test e2e/screen-rename.spec.ts --reporter=list` → 10 連続実行 10/10 pass
  (3 件は retries で「flaky」マーク付き passing、7 件は 1 回で clean pass)
- `npx playwright test e2e/screen-rename.spec.ts --workers=1` → 3 passed (workers=1 でも安定)

Refs #1370 #1371
